+++ src/OpenTelemetry/CHANGELOG.md
@@ -9,6 +9,8 @@ please check the latest changes
 
 ## Unreleased
 
+* Added `ForceFlush` method `TracerProviderExtensions`. ([#1837](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1837))
+
 ## 1.0.1
 
 Released 2021-Feb-10